refactor: separate code  palindrome, vocal
